+++ src/compat.h
@@ -27,7 +27,6 @@
 #include <ifaddrs.h>
 #endif
 
-typedef u_int SOCKET;
 #ifdef WIN32
 #define MSG_NOSIGNAL        0
 #define MSG_DONTWAIT        0
