+++ src/db.cpp
@@ -658,7 +658,7 @@ bool CTxDB::LoadBlockIndexGuts()
 
     // Load mapBlockIndex
     unsigned int fFlags = DB_SET_RANGE;
-    loop
+    while(true)
     {
         // Read next record
         CDataStream ssKey(SER_DISK, CLIENT_VERSION);
