+++ src/irc.cpp
@@ -79,7 +79,7 @@ static bool Send(SOCKET hSocket, const char* pszSend)
 
 bool RecvLineIRC(SOCKET hSocket, string& strLine)
 {
-    loop
+    while(true)
     {
         bool fRet = RecvLine(hSocket, strLine);
         if (fRet)
@@ -102,7 +102,7 @@ bool RecvLineIRC(SOCKET hSocket, string& strLine)
 
 int RecvUntil(SOCKET hSocket, const char* psz1, const char* psz2=NULL, const char* psz3=NULL, const char* psz4=NULL)
 {
-    loop
+    while(true)
     {
         string strLine;
         strLine.reserve(10000);
@@ -137,7 +137,7 @@ bool Wait(int nSeconds)
 bool RecvCodeLine(SOCKET hSocket, const char* psz1, string& strRet)
 {
     strRet.clear();
-    loop
+    while(true)
     {
         string strLine;
         if (!RecvLineIRC(hSocket, strLine))
