+++ src/qt/bitcoingui.cpp
@@ -20,15 +20,17 @@
 #include "addresstablemodel.h"
 #include "transactionview.h"
 #include "overviewpage.h"
-#include "miningpage.h"
 #include "bitcoinunits.h"
 #include "guiconstants.h"
 #include "askpassphrasedialog.h"
 #include "notificator.h"
 #include "guiutil.h"
 #include "rpcconsole.h"
+#include "wallet.h"
 #include "blockbrowser.h"
-#ifdef Q_WS_MAC
+#include "miningpage.h"
+
+#ifdef Q_OS_MAC
 #include "macdockiconhandler.h"
 #endif
 
@@ -46,6 +48,7 @@
 #include <QPushButton>
 #include <QLocale>
 #include <QMessageBox>
+#include <QMimeData>
 #include <QProgressBar>
 #include <QStackedWidget>
 #include <QDateTime>
@@ -55,11 +58,9 @@
 #include <QTimer>
 #include <QDragEnterEvent>
 #include <QUrl>
-#include <QFile>
-#include <QTextStream>
+#include <QStyle>
 #include <QSignalMapper>
-#include <QSettings>
-#include <iostream>
+#include <QTextStream>
 
 BitcoinGUI::BitcoinGUI(QWidget *parent):
     QMainWindow(parent),
