+++ src/util.cpp
@@ -280,7 +280,7 @@ string vstrprintf(const std::string &format, va_list ap)
     char* p = buffer;
     int limit = sizeof(buffer);
     int ret;
-    loop
+    while(true)
     {
         va_list arg_ptr;
         va_copy(arg_ptr, ap);
@@ -327,7 +327,7 @@ void ParseString(const string& str, char c, vector<string>& v)
         return;
     string::size_type i1 = 0;
     string::size_type i2;
-    loop
+    while(true)
     {
         i2 = str.find(c, i1);
         if (i2 == str.npos)
@@ -443,7 +443,7 @@ vector<unsigned char> ParseHex(const char* psz)
 {
     // convert hex dump to vector
     vector<unsigned char> vch;
-    loop
+    while(true)
     {
         while (isspace(*psz))
             psz++;
@@ -897,7 +897,7 @@ string DecodeBase32(const string& str)
 
 bool WildcardMatch(const char* psz, const char* mask)
 {
-    loop
+    while(true)
     {
         switch (*mask)
         {
